+++ frontend/src/components/VolunteerDates.js
@@ -40,6 +40,14 @@ export default function VolunteerDates({
         if (checked) {
             selectedDateIDs.push(date.id);
             setSelectedDateIDs(selectedDateIDs);
+            
+            for (let i = 0; i < unselectedDateIDs.length; i++) {
+                if (unselectedDateIDs[i] === date.id) {
+                    unselectedDateIDs.splice(i, 1);
+                    break;
+                }
+            }
+
         } else {
             for (let i = 0; i < selectedDateIDs.length; i++) {
                 if (selectedDateIDs[i] === date.id) {
@@ -56,6 +64,9 @@ export default function VolunteerDates({
                 setUnselectedDateIDs(unselectedDateIDs);
             }
         }
+
+        console.log(selectedDateIDs)
+        console.log(unselectedDateIDs)
     };
 
     // It's hacky but it works
