+++ frontend/src/pages/HomePage.js
@@ -1,6 +1,5 @@
 import Button from "react-bootstrap/Button";
 import React, { useState, useEffect } from "react";
-import VolunteerDates from "../components/VolunteerDates";
 import { useNavigate } from "react-router-dom";
 import {
     signOutUser,
@@ -9,16 +8,16 @@ import {
     getSignedInUserInfoFB,
 } from "../FirebaseFunctions";
 import { useAuthState } from "react-firebase-hooks/auth";
+import SignUpPopup from "../components/SignUpPopup";
 
 export default function HomePage() {
-    const [selectedDateIDs, setSelectedDateIDs] = useState([]);
-    const [unselectedDateIDs, setUnselectedDateIDs] = useState([]);
-    const [availableDates, setAvailableDates] = useState([]);
     const [username, setUsername] = useState("");
-    const [fullUserName, setFullUsername] = useState(null);
     const [user, loading] = useAuthState(auth);
 
     const navigate = useNavigate();
+
+    const [showSignUpPopup, setShowSignUpPopup] = useState(false);
+    const [fullUserName, setFullUsername] = useState(null);
 
     useEffect(() => {
         const getUsername = async () => {
@@ -32,41 +31,12 @@ export default function HomePage() {
                 setUsername(info.name.split(" ")[0]);
                 setFullUsername(info.name);
             } catch {
-                 alert("You must be signed in to view this page");
+                alert("You must be signed in to view this page");
                 window.location.href = "/login";
             }
         };
         getUsername();
     }, []);
-
-    const submit = async (e) => {
-        e.preventDefault();
-
-        const response = await getSignedInUserInfoFB();
-
-        if (!response.watchedVideo || !response.uploadedForm) {
-            alert(
-                "You must watch the video and upload the form before signing up for a date!"
-            );
-            return;
-        }
-
-        let request = "Error";
-        try {
-            request = await changeDateFB(selectedDateIDs, unselectedDateIDs);
-        } catch {
-            alert("There was an error with your request");
-        }
-
-        if (request.data === "User already Signed Up")
-            alert("You are already signed up for this date");
-        else if (request.data === "No date found with that ID.")
-            alert("There was an error with your request");
-        else if (request.data === "No user signed in")
-            alert("You must be signed in to sign up for a date");
-        else if (request.data === "success")
-            alert("Your date changes have been saved!");
-    };
 
     const handleLogout = async () => {
         try {
@@ -84,10 +54,10 @@ export default function HomePage() {
                 className="fixed-top navbar NavHead"
                 style={{ textAlign: "center" }}
             >
-                <h1 style={{}} className="CPStyleFull">
+                <h2 style={{}} className="CPStyleFull">
                     Thank you for volunteering for the Champion Project{" "}
                     {username}!
-                </h1>
+                </h2>
                 <h1 style={{}} className="CPStyleMobile">
                     Hi {username}!
                 </h1>
@@ -152,7 +122,25 @@ export default function HomePage() {
                         Change volunteer event dates and view volunteers
                     </p>
                 </div>
+                <div className="Statistic">
+                    <button
+                        type="button"
+                        className="btn btn-primary btn-square-md"
+                        onClick={() => setShowSignUpPopup(true)}
+                    >
+                        Sign Up
+                    </button>
+                    <p className="StatDescription">
+                        Sign up to volunteer at a bookbus event
+                    </p>
+                </div>
             </div>
+
+            <SignUpPopup
+                showSignUpPopup={showSignUpPopup}
+                setShowSignUpPopup={setShowSignUpPopup}
+                fullUserName={fullUserName}
+            />
         </>
     );
 }
